+++ SharpCommunicator/src/main/java/ch/erzberger/sharppc/ProcessFiles.java
@@ -22,7 +22,7 @@ public class ProcessFiles implements ByteProcessor {
 
     public ProcessFiles(String filename, PocketPcDevice device) {
         this.filename = filename;
-        if (PocketPcDevice.PC1500.equals(device)) {
+        if (device.isPC1500()) {
             // Through trial and error, it looks line 3 Seconds is enough to write even a long line
             timeout = 3000L;
         } else {